code_review_backend/requirements.txt
# FastAPI и веб-сервер
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<0.31
pydantic>=2.9,<3

# HTTP клиент для GitLab API
requests>=2.32,<3

# Gemini API
google-generativeai>=0.8.0,<1

# Конфигурация
python-dotenv>=1.0,<2
